Updated changelog and added deprecation errors to DistributedInstruments
