+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-...
+### Changed
+- Moved InstrumentDataServer to qtt.instrument_drivers.instrument_data_server (#690)
+- Moved InstrumentDataClient to qtt.instrument_drivers.instrument_data_client (#690) 
+- Moved BlueforsApp to qtt.instrument_drivers.BlueforsMonitorApp (#690)
+
+### Deprecated
+- InstrumentDataServer removed from qtt.instrument_drivers.DistributedInstrument (#690)
+- InstrumentDataClient removed from qtt.instrument_drivers.DistributedInstrument (#690) 
 
 ## [1.1.3] - 2019-10-18
